+++ src/requestBody/TeacherRegisterBody.kt
@@ -4,5 +4,5 @@ data class TeacherRegisterBody (
     var name: String,
     var password: String,
     var teachingSpecialization: String,
-    var phone: Int,
+    var phone: Long,
 )
